requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
pydantic==1.10.9
numpy==1.26.4
scikit-learn==1.4.2
networkx==3.2
prometheus-client==0.17.0
jsonschema==4.19.0
faiss-cpu==1.7.4
pytest==7.4.2
pytest-asyncio==0.22.0
python-multipart==0.0.6